fastapi-chat/requirements.txt
fastapi == 0.46.0
uvicorn == 0.11.1
requests == 2.22.0
firebase-admin ~= 1.0.5
pydantic ~= 1.9.3
python-multipart ~= 0.0.5 